feat: init

+++ app/Model/GamePlayer.php
@@ -16,9 +16,9 @@ use Hyperf\Database\Model\Relations\BelongsTo;
  * @property int $seat
  * @property string $name
  * @property bool $is_hero
- * @property float $stack 初始筹码
+ * @property int $stack 初始筹码
  * @property SeatTypeEnum $seat_type 座位类型
- * @property ?float $blind_amount 盲注
+ * @property ?int $blind_amount 盲注
  * @property list<string>|null $cards 已知手牌
  */
 class GamePlayer extends Model
@@ -31,8 +31,8 @@ class GamePlayer extends Model
         'seat_type' => SeatTypeEnum::class,
         'cards' => 'array',
         'is_hero' => 'boolean',
-        'stack' => 'float',
-        'blind_amount' => 'float',
+        'stack' => 'integer',
+        'blind_amount' => 'integer',
     ];
 
     /** @return BelongsTo<Game, static> */
